+++ CHANGELOG.md
@@ -8,6 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Custom MT Saltshaker HTML report loading (#6242)
 - Command to download all PanelApp panels to file (#6239)
+- Alternatively load and update PANELAPP-GREEN panel from a pre-downloaded file (#6244)
 
 ## [4.110.0]
 ### Added